Fixed layout.tsx and page.tsx, moved queries to correct file, updated fetchRegisteredEvents logic
